repo-bisect: do proper bisect
The bisect code was a bit confused, and would always look at the versions
at the boundaries, instead of the versions in the middle, thus not doing
a proper bisect.

This change fixes this behaviour, the revision tested will be in the middle
of the current range, and will then become the lower (for a good revision) or
upper (for a bad revision) boundary for the next iteration.